requirements.txt
# Core dependencies for production
click>=8.0.0
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.28.0
